move class User to package models
